+++ lib/speed_up_get.dart
@@ -1,7 +1,4 @@
 library speed_up_get;
 
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}
+export './src/extensions.dart';
+export './src/mixins.dart';