+++ lib/src/extensions.dart
@@ -1,0 +1,9 @@
+import 'package:get/get.dart';
+
+extension GetxViewExtensions<T> on GetView<T> {
+  T get c => controller;
+}
+
+extension GetxWidgetExtensions<T extends GetLifeCycleBase?> on GetWidget<T> {
+  T get c => controller;
+}